update sgm to have clear interface to sgm_Carbon and sgm_Nitro
Major redesign of SGM to accommodate the VIS adn SWIR radiation scene modules. This requires a scene generator for the geohysical scene and a VIS and SWIR radiation scene generator. New interfaces are introduced. Also several S2 band can be retrieved, e.g., B02 and B03 for the VIS simulations  and B11 for the SWIR simulations. The SWIR radiation scene generator is updated as well.