Allow to build L1B and IM as a single project
There is now a top-level CMakeLists.txt that builds both the C++ IM and
L1B as a single project. Once the tests are present, there will be
several directories that depend on each other. If each directory was a
separate CMake project it would complicate the detection of
dependencies. Having a top-level wrapper to bind everything into a
single project simplifies the configuration and build process.

The L1B processor can still be built independently if required.