Refactor IO operations
Move reading and writing functions to io.py files of the respective
modules. For instance, all read/write functions related to geometry are
in teds.gm.io. This facilitates reuse and reduces clutter.

Do the same for types. For instance, Geometry and Navigation are now in
teds.gm.types. Previously, they were in teds.l1al1b.python.types to
avoid a circular import error but this got fixed.