Restructure Python and C++ into separate folders
Move all C++ source files out of the teds directory. Mixing the
two languages led to unnecessary complications when dealing with imports
and is in general not considered good style.

The C++ extensions are now built using the same CMake build scripts as
the operational build system. See changes in setup.py.