Add support for Stokes Q,U and Mueller elements
In SGM, if [io_files][q_u] points to a file, then the Stokes q and u are
read from there and the Stokes Q and U are written into the radiance
product. If the radiance is convolved with the ISRF in the SGM then Q
and U are also convolved before saving to output.

In IM, if [io_files][mueller] points to a file, then the Mueller matrix
elements are read from there and the radiance is modified by adding the
polarization contribution
  I_pol = I + M_01 / M_00 * Q + M_02 / M_00 * U.